# openapi.yml
openapi: 3.0.0
info:
  title: Distributed Raft Key-Value Store API
  description: |
    A fault-tolerant, strongly consistent key-value store implemented using the Raft consensus algorithm.
    The API provides operations for setting and getting values in a distributed cluster while maintaining
    data consistency across node failures.
    
    ## Key Features
    - **Strong Consistency**: All reads and writes are linearizable
    - **Fault Tolerance**: Survives up to (N-1)/2 node failures in an N-node cluster
    - **Automatic Failover**: Leader election ensures continued availability
    - **No Data Loss**: All committed data persists across crashes
    
    ## Cluster Information
    - **Leader-based Model**: Only the leader accepts writes
    - **Replication**: Data is replicated to followers with majority-based commitment
    - **Raft Consensus**: Custom gRPC-based Raft implementation
    - **Monitoring**: Prometheus metrics exposed for cluster health monitoring
    
    ## Authentication
    Currently, no authentication is implemented. All endpoints are publicly accessible.
    For production deployments, implement TLS and authentication layers.
  version: 1.0.0
  contact:
    name: Distributed KV Store Team
    url: https://github.com/Ashwanthbingi/raft-harmony
  license:
    name: MIT
servers:
  - url: http://localhost:8001
    description: Local development (Node 1)
  - url: http://localhost:8002
    description: Local development (Node 2)
...
paths:
  /set:
    post:
      tags:
        - Key-Value Operations
      summary: Set a key-value pair
      description: |
        Store or update a key-value pair in the distributed store.
        
        The request is processed by the **leader node only**. If sent to a follower, 
        the follower will redirect or reject with a message indicating the current leader.
        
        The operation is completed when:
        1. The entry is appended to the leader's log
        2. The leader replicates it to followers
        3. A majority of nodes acknowledge the entry
        4. The leader commits and applies the entry to the state machine
        
        **Consistency Guarantee**: Once the operation returns successfully, all subsequent
        reads will see the written value (strong consistency/linearizability).
      operationId: setValue
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - key
                - value
              properties:
                key:
                  type: string
                  description: The key to store
                  example: "username"
                  minLength: 1
                  maxLength: 256
                value:
                  type: string
                  description: The value to store
                  example: "alice"
                  minLength: 0
                  maxLength: 1048576
          application/json:
            schema:
              type: object
              required:
                - key
                - value
              properties:
                key:
                  type: string
                  description: The key to store
                  example: "username"
                value:
                  type: string
                  description: The value to store
                  example: "alice"
      parameters:
        - name: key
          in: query
          description: The key to store
          schema:
            type: string
          example: "username"
        - name: value
          in: query
          description: The value to store
          schema:
            type: string
          example: "alice"
      responses:
        '200':
          description: Value successfully submitted to the Raft log
          content:
            text/plain:
              schema:
                type: string
              example: "Submitted at index 42 term 3"
        '400':
          description: Missing required parameters (key or value)
          content:
            text/plain:
              schema:
                type: string
              example: "Missing key/value"
        '500':
          description: |
            Node is not the leader or server error.
            Try another node or wait for leader election to complete.
          content:
            text/plain:
              schema:
                type: string
              example: "Not Leader. Leader is node2"
      x-code-samples:
        - lang: curl
          source: |
            curl -X POST "http://localhost:8001/set?key=username&value=alice"
        - lang: javascript
          source: |
            fetch('http://localhost:8001/set', {
              method: 'POST',
              headers: { 'Content-Type': 'application/json' },
              body: JSON.stringify({ key: 'username', value: 'alice' })
            })
            .then(r => r.text())
            .then(console.log)
  /get:
    get:
      tags:
        - Key-Value Operations
      summary: Get a value by key
      description: |
        Retrieve the value associated with a key from the state machine.
        
        This operation reads from the **applied state machine**. The read reflects
        all values that have been committed by the Raft cluster. There are no consistency
        guarantees about this specific node's view (use /health for leader verification).
        
        **Data Freshness**: 
        - On the leader: Current data (up to last applied index)
        - On a follower: May lag behind the leader by up to one replication round
        
        For strict linearizable reads, send requests to the leader node only.
      operationId: getValue
      parameters:
        - name: key
          in: query
          required: true
          description: The key to retrieve
          schema:
            type: string
          example: "username"
      responses:
        '200':
          description: Key found; value returned
          content:
            text/plain:
              schema:
                type: string
              example: "alice"
        '404':
          description: Key not found in the state machine
          content:
            text/plain:
              schema:
                type: string
              example: "Not found"
        '400':
          description: Missing required key parameter
          content:
            text/plain:
              schema:
                type: string
              example: "Missing key parameter"
      x-code-samples:
        - lang: curl
          source: |
            curl "http://localhost:8001/get?key=username"
        - lang: javascript
          source: |
            fetch('http://localhost:8001/get?key=username')
              .then(r => r.text())
              .then(console.log)
...
  /range:
    get:
      tags:
        - Range Queries
      summary: Get keys in a range
      description: |
        Retrieve all key-value pairs where keys fall within the specified range [start, end).
        
        Useful for pagination, range scans, and ordered queries.
        
        **Example Use Cases:**
        - Get scores between 1000-2000: /range?start=score:1000&end=score:2000
        - Get entries after a timestamp: /range?start=2024-07-10T12:00:00
      operationId: getRangeKeys
      parameters:
        - name: start
          in: query
          required: true
          description: Starting key (inclusive)
          schema:
            type: string
          example: "user:100"
        - name: end
          in: query
          required: false
          description: Ending key (exclusive). If omitted, no upper bound.
          schema:
            type: string
          example: "user:200"
      responses:
        '200':
          description: Keys in range retrieved successfully
          content:
            application/json:
             schema:
               type: object
               properties:
                 keys:
                   type: object
                   additionalProperties:
                     type: string
                 count:
                   type: integer
                 truncated:
                   type: boolean
                 range:
                   type: object
                   properties:
                     start:
                       type: string
                     end:
                       type: string
        '400':
          description: Missing required parameters
      x-code-samples:
        - lang: curl
          source: |
            curl "http://localhost:8001/range?start=user:100&end=user:200"
  /prefix:
    get:
      tags:
        - Range Queries
      summary: Get keys by prefix
      description: |
        Retrieve all key-value pairs that start with the specified prefix.
        
        Efficient for hierarchical key organization.
        
        **Example Use Cases:**
        - Get all user settings: /prefix?prefix=user:settings:
        - Get all config values: /prefix?prefix=config:
        - Get all cache entries: /prefix?prefix=cache:session:
      operationId: getPrefixKeys
      parameters:
        - name: prefix
          in: query
          required: true
          description: Key prefix to match
          schema:
            type: string
          example: "user:"
      responses:
        '200':
          description: Keys with prefix retrieved successfully
          content:
            application/json:
             schema:
               type: object
               properties:
                 keys:
                   type: object
                   additionalProperties:
                     type: string
                 count:
                   type: integer
                 prefix:
                   type: string
                 truncated:
                   type: boolean
        '400':
          description: Missing prefix parameter
      x-code-samples:
        - lang: curl
          source: |
            curl "http://localhost:8001/prefix?prefix=user:"
  /keys:
    get:
      tags:
        - Range Queries
      summary: List all keys with pagination
      description: |
        Retrieve all keys in the store with optional pagination.
        
        Keys are returned in sorted order.
        
        **Pagination:**
        - Use limit and offset for efficient pagination
        - Default limit: 100 keys per page
        - Check hasMore field to determine if more keys exist
      operationId: listKeys
      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum keys to return (default 100)
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          required: false
          description: Pagination offset (default 0)
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Keys listed successfully
          content:
            application/json:
             schema:
               type: object
               properties:
                 keys:
                   type: array
                   items:
                     type: string
                 count:
                   type: integer
                 total:
                   type: integer
                 offset:
                   type: integer
                 limit:
                   type: integer
                 truncated:
                   type: boolean
                 hasMore:
                   type: boolean
      x-code-samples:
        - lang: curl
          source: |
            curl "http://localhost:8001/keys?limit=50&offset=0"
  /batch/get:
    post:
      tags:
        - Batch Operations
      summary: Get multiple keys at once
      description: |
        Retrieve values for multiple keys in a single request.
        
        More efficient than individual GET requests for retrieving many keys.
        
        **Returns:**
        - Only keys that exist in the store
        - Use the "found" and "missing" counts to verify completeness
      operationId: batchGet
      requestBody:
        required: true
        content:
          application/json:
            schema:
             type: object
             required:
               - keys
             properties:
               keys:
                 type: array
                 items:
                   type: string
                 description: List of keys to retrieve
             example:
               keys: ["user:1", "user:2", "user:3"]
      responses:
        '200':
          description: Batch GET completed
          content:
            application/json:
             schema:
               type: object
               properties:
                 results:
                   type: object
                   additionalProperties:
                     type: string
                 count:
                   type: integer
                 found:
                   type: integer
                 missing:
                   type: integer
        '400':
          description: Invalid request or empty keys list
      x-code-samples:
        - lang: curl
          source: |
            curl -X POST http://localhost:8001/batch/get \
             -H "Content-Type: application/json" \
             -d '{"keys": ["user:1", "user:2", "user:3"]}'
  /batch/set:
    post:
      tags:
        - Batch Operations
      summary: Set multiple key-value pairs
      description: |
        Set multiple key-value pairs in the distributed store.
        
        Each SET operation is individually replicated via Raft.
        Batch is not atomic (partial success possible).
        
        **Note:** Only available on the leader node.
      operationId: batchSet
      requestBody:
        required: true
        content:
          application/json:
            schema:
             type: object
             required:
               - operations
             properties:
               operations:
                 type: array
                 items:
                   type: object
                   required:
                     - key
                     - value
                   properties:
                     key:
                       type: string
                     value:
                       type: string
             example:
               operations:
                 - key: "user:1"
                   value: "Alice"
                 - key: "user:2"
                   value: "Bob"
      responses:
        '200':
          description: Batch SET completed
          content:
            application/json:
             schema:
               type: object
               properties:
                 results:
                   type: object
                   additionalProperties:
                     type: object
                     properties:
                       success:
                         type: boolean
                       index:
                         type: integer
                       term:
                         type: integer
                       error:
                         type: string
                 total:
                   type: integer
                 success:
                   type: integer
                 failed:
                   type: integer
        '400':
          description: Invalid request or empty operations
        '500':
          description: Only leader can perform SET operations
      x-code-samples:
        - lang: curl
          source: |
            curl -X POST http://localhost:8001/batch/set \
             -H "Content-Type: application/json" \
             -d '{
               "operations": [
                 {"key": "user:1", "value": "Alice"},
                 {"key": "user:2", "value": "Bob"}
               ]
             }'
  /batch/delete:
    post:
      tags:
        - Batch Operations
      summary: Delete multiple keys
      description: |
        Delete multiple keys from the distributed store.
        
        Each DELETE operation is individually replicated via Raft.
        
        **Note:** Only available on the leader node.
      operationId: batchDelete
      requestBody:
        required: true
        content:
          application/json:
            schema:
             type: object
             required:
               - keys
             properties:
               keys:
                 type: array
                 items:
                   type: string
             example:
               keys: ["user:1", "user:2", "user:3"]
      responses:
        '200':
          description: Batch DELETE completed
          content:
            application/json:
             schema:
               type: object
               properties:
                 results:
                   type: object
                   additionalProperties:
                     type: object
                     properties:
                       success:
                         type: boolean
                       index:
                         type: integer
                       term:
                         type: integer
                 total:
                   type: integer
                 success:
                   type: integer
                 failed:
                   type: integer
        '400':
          description: Invalid request or empty keys
        '500':
          description: Only leader can perform DELETE operations
  /add-peer:
    post:
      tags:
        - Cluster Management
      summary: Add a peer to the cluster
      description: |
        Add a new node to the Raft cluster.
        
        **Note:** Only the leader can add peers.
      operationId: addPeer
      parameters:
        - name: id
          in: query
          required: true
          description: Node ID of the new peer
          schema:
            type: string
          example: "node4"
        - name: addr
          in: query
          required: true
          description: Network address of the new peer (host:port)
          schema:
            type: string
          example: "10.0.4.10:9000"
      responses:
        '200':
          description: Peer added successfully
        '400':
          description: Missing id or addr parameter
        '500':
          description: Only leader can add peers
  /remove-peer:
    post:
      tags:
        - Cluster Management
      summary: Remove a peer from the cluster
      description: |
        Remove a node from the Raft cluster.
        
        **Note:** Only the leader can remove peers.
      operationId: removePeer
      parameters:
        - name: id
          in: query
          required: true
          description: Node ID to remove
          schema:
            type: string
          example: "node4"
      responses:
        '200':
          description: Peer removed successfully
        '400':
          description: Missing id parameter
        '500':
          description: Only leader can remove peers
...